--- engine/language-server/src/main/scala/org/enso/languageserver/http/server/BinaryEncoder.scala
@@ -1,9 +1,0 @@
-package org.enso.languageserver.http.server
-
-import java.nio.ByteBuffer
-
-trait BinaryEncoder[-A] {
-
-  def encode(msg: A): ByteBuffer
-
-}
