--- engine/language-server/src/main/scala/org/enso/languageserver/protocol/binary/BinaryProtocolEncoder.scala
@@ -1,11 +1,0 @@
-package org.enso.languageserver.protocol.binary
-
-import java.nio.ByteBuffer
-
-import org.enso.languageserver.http.server.BinaryEncoder
-
-object BinaryProtocolEncoder extends BinaryEncoder[ByteBuffer] {
-
-  override def encode(msg: ByteBuffer): ByteBuffer = identity(msg)
-
-}
