+++ engine/language-server/src/main/scala/org/enso/languageserver/util/binary/BinaryEncoder.scala
@@ -1,0 +1,15 @@
+package org.enso.languageserver.util.binary
+
+import java.nio.ByteBuffer
+
+trait BinaryEncoder[-A] {
+
+  def encode(msg: A): ByteBuffer
+
+}
+
+object BinaryEncoder {
+
+  val empty: BinaryEncoder[ByteBuffer] = identity
+
+}
