+++ engine/language-server/src/test/scala/org/enso/languageserver/websocket/binary/OutboundMessageDecoder.scala
@@ -2,8 +2,8 @@ package org.enso.languageserver.websocket.binary
 
 import java.nio.ByteBuffer
 
-import org.enso.languageserver.http.server.BinaryDecoder
 import org.enso.languageserver.protocol.binary.envelope.OutboundMessage
+import org.enso.languageserver.util.binary.BinaryDecoder
 
 object OutboundMessageDecoder extends BinaryDecoder[OutboundMessage] {
   override def decode(bytes: ByteBuffer): OutboundMessage = {
